requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alkivi-logger==1.2.5
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
click==6.7
idna==2.6                 # via requests
requests==2.18.4          # via slackbot, slacker
six==1.11.0               # via slackbot, websocket-client
slackbot==0.5.1
slacker==0.9.60           # via slackbot
urllib3==1.22             # via requests
websocket-client==0.44.0  # via slackbot
